- appearance rewrite - fixed flaw in button clicking packet - fixed flaw in appearance change packet

+++ src/main/java/com/cryptic/model/entity/player/save/PlayerSave.java
@@ -509,7 +509,7 @@ public class PlayerSave {
         private final Item topPkerReward;
         private final boolean female;
         private final int[] looks;
-        private final int[] colors;
+        private final byte[] colors;
 
         //Combat attribs
         private final String spellBook;
